fix the issue where mp4 cannot be converted to mkv
when converting an MP4 file to an MKV file,the tags are incompatible
with the output codec ID.
because mp4a is the tag that marks the AAC audio stream in an MP4
file,the MKV container is not compatible with this tag.
set the codec_tag of the audio stream to 0 to avoid compatibility
issues.